Field Ids are shorts now per feature #14

+++ java/src/main/java/com/imprint/core/DirectoryEntry.java
@@ -7,17 +7,17 @@ import java.util.Objects;
 
 /**
  * A directory entry describing a single field in an Imprint record.
- * Each entry has a fixed size of 9 bytes.
+ * Each entry has a fixed size of 7 bytes.
  */
 @Getter
 @EqualsAndHashCode
 public final class DirectoryEntry {
-    private final int id;
+    private final short id;
     private final TypeCode typeCode;
     private final int offset;
     
     public DirectoryEntry(int id, TypeCode typeCode, int offset) {
-        this.id = id;
+        this.id = (short) id;
         this.typeCode = Objects.requireNonNull(typeCode, "TypeCode cannot be null");
         this.offset = offset;
     }
